ssh-pregen-hostkeys: Limit to qemu machines by default
There are potential security issues from using pre-generated host keys. We made
the recipe available for autobuilder testing purposes but concerns remain about
how easily this could end up in production.

I thought we'd already done this, but limit the recipe to qemu* machines,
which means any real hardware trying to use it will need to be a bit more
explicit about it and specifically enable it.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>